Adição de switch no formulario de login e cadastro
