+++ front/src/components/admin/Layout/AdminLayout.css
@@ -1,0 +1,74 @@
+@import url("https://cdn.jsdelivr.net/npm/bootstrap-icons@1.13.1/font/bootstrap-icons.min.css");
+
+.bodyAdmin {
+  padding: 1rem 1rem 1rem 0;
+  background-color: white;
+}
+
+.mainAdmin {
+  display: flex;
+  gap: 1rem;
+}
+
+/* nav padrão admin*/
+.navAdmin {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 95vh;
+  width: 18rem;
+  padding: 1rem 0;
+  align-items: center;
+  box-sizing: border-box;
+}
+
+.links-nav {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+  width: 100%;
+  align-items: center;
+}
+
+.links-nav a {
+  text-decoration: none;
+  color: #3498db;
+  border: 1px solid #3498db;
+  width: 7rem;
+  padding: 0.8rem 1.9rem;
+  border-radius: 8px;
+  display: flex;
+  justify-content: flex-start;
+  gap: 0.5rem;
+}
+
+.links-nav a:hover {
+  background-color: #3498db;
+  color: white;
+}
+
+.links-nav a.active {
+  background-color: #3498db; 
+  color: white; 
+}
+
+/* area de conteudo padrão de admin */
+.fundo-container-admin {
+  display: flex;
+  flex: 1;
+  background-color: rgb(81, 108, 202, 0.1);
+  height: 95vh;
+  box-sizing: border-box;
+  border-radius: 8px;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.05);
+  padding: 2rem;
+  overflow-y: auto; 
+}
+
+.fundo-content-admin {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start; 
+}
