+++ front/src/components/user/Cadastrar/Cadastro.css
@@ -103,13 +103,31 @@
 }
 
 /* Estilo do formulário */
+/* Força o formulário a ser uma coluna e não uma linha */
 .form-cadastro {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    max-width: 400px; /* Evita que os inputs fiquem gigantes horizontalmente */
+    gap: 10px; /* Espaçamento entre os inputs */
+}
+
+.input-cadastro {
+    width: 100%;
+    box-sizing: border-box; /* Garante que o padding não aumente a largura */
+    padding: 10px;
+    border: 2px solid #3498DB;
+    border-radius: 8px;
+    outline: none;
+}
+
+/* Garante que o container interno do cadastro não herde comportamentos estranhos */
+.signup-right-panel-inner {
     display: flex;
     flex-direction: column;
     align-items: center;
     width: 100%;
 }
-
 /*.label-cadastro {
     width: 90%;
     font-size: 12px;
@@ -132,23 +150,22 @@
 }
 
 .button-cadastro {
-  width: 35%; /* Ajustado para 90% do form, menos 20px de padding horizontal */
-  padding: 13px; /* Seu padding */
-  background-color: #3498DB; /* Sua cor */
+  width: 52%;
+  padding: 11px;
+  background-color: #3498db;
   color: #fff;
   border: none;
-  border-radius: 6px; /* Seu border-radius */
-  font-size: 20px; /* Seu font-size */
+  border-radius: 6px;
+  font-size: 15px;
   font-weight: bold;
   cursor: pointer;
-  margin-top: 10px; /* Seu margin-top */
-  box-sizing: border-box; /* Garante que padding não adicione à largura total */
+  box-sizing: border-box;
   transition: background-color 0.3s ease;
-  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.3);
+  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.25);
 }
 
 .button-cadastro:hover {
-    background-color: #196292;
+    background-color: #1972b5;
 }
 
 /* Estilo do logo UFMS no rodapé */
