+++ front/src/components/user/Login/Logar.css
@@ -29,9 +29,12 @@
 }
 
 .login-container::before {
-  content: '';
+  content: "";
   position: absolute;
-  top: -10px; left: -10px; right: -10px; bottom: -10px;
+  top: -10px;
+  left: -10px;
+  right: -10px;
+  bottom: -10px;
   background: linear-gradient(to top, #2c7bb1, #4da4e4, #66b6eb, #a0dfff);
   z-index: -1;
   filter: blur(20px);
@@ -41,7 +44,7 @@
 /* ---- Divisor ---- */
 .login-divider {
   width: 1.2px;
-  background-color: #3498DB;
+  background-color: #3498db;
   position: absolute;
   left: 50%;
   transform: translateX(-50%);
@@ -87,15 +90,14 @@
   width: 50%;
   display: flex;
   flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 36px 28px;
+  align-items: center; /* Centraliza o switch e o form */
+  justify-content: flex-start; /* Alinha o conteúdo no topo para o switch ficar em cima */
+  padding: 40px 20px;
   box-sizing: border-box;
-  text-align: center;
 }
 
 .login-title {
-  color: #3498DB;
+  color: #3498db;
   font-size: 2.4em;
   font-weight: bold;
   margin: 0 0 28px 0;
@@ -150,7 +152,7 @@
 
 .login-input:focus {
   border-color: #1a7fc4;
-  box-shadow: 0 0 0 3px rgba(52,152,219,0.15);
+  box-shadow: 0 0 0 3px rgba(52, 152, 219, 0.15);
 }
 
 /* ---- Esqueci minha senha ---- */
@@ -171,7 +173,7 @@
 .login-btn-submit {
   width: 52%;
   padding: 11px;
-  background-color: #3498DB;
+  background-color: #3498db;
   color: #fff;
   border: none;
   border-radius: 6px;
@@ -180,7 +182,7 @@
   cursor: pointer;
   box-sizing: border-box;
   transition: background-color 0.3s ease;
-  text-shadow: 1px 1px 2px rgba(0,0,0,0.25);
+  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.25);
 }
 
 .login-btn-submit:hover {
@@ -196,7 +198,7 @@
 }
 
 .login-register-link {
-  color: #3498DB;
+  color: #3498db;
   text-decoration: none;
   font-weight: bold;
 }
@@ -248,7 +250,9 @@
     min-height: unset;
   }
 
-  .login-divider { display: none; }
+  .login-divider {
+    display: none;
+  }
 
   .login-left-panel,
   .login-right-panel {
@@ -256,19 +260,31 @@
     padding: 24px 20px;
   }
 
-  .login-logo { width: 130px; }
+  .login-logo {
+    width: 130px;
+  }
 
   .login-title {
     font-size: 1.8em;
     margin-bottom: 16px;
   }
 
-  .login-btn-submit { width: 65%; }
+  .login-btn-submit {
+    width: 65%;
+  }
 }
 
 @media (max-width: 480px) {
-  .login-title { font-size: 1.5em; }
-  .login-input { font-size: 13px; }
-  .login-btn-submit { width: 75%; }
-  .login-logo { width: 110px; }
+  .login-title {
+    font-size: 1.5em;
+  }
+  .login-input {
+    font-size: 13px;
+  }
+  .login-btn-submit {
+    width: 75%;
+  }
+  .login-logo {
+    width: 110px;
+  }
 }
